kshay's code cleanup (no kernel/futex.c)
Former-commit-id: 488254ae614225f7f69c3fe13f6f1da2820b11f0

+++ fs/proc/krninfo.c
@@ -1,3 +1,5 @@
+/* (C) 2014, Akshay Ravichandran, SSRG Virginia Tech */
+
 #include <linux/fs.h>
 #include <linux/init.h>
 #include <linux/proc_fs.h>
@@ -7,13 +9,12 @@
 
 static int krninfo_proc_show(struct seq_file *m, void *v)
 {
-
 	seq_printf(m,
 		"POPCORN KERNEL INFO:   \n"
 		"Kernel Id    %8lu \n"
 		"Start PFN    %8lu \n"
 		"End   PFN    %8lu \n",
-		Kernel_Id,kernel_start_addr,PFN_PHYS(max_low_pfn));
+		(unsigned long) Kernel_Id, kernel_start_addr, (unsigned long) PFN_PHYS(max_low_pfn));
 
 	return 0;
 }
